+++ zerver/tests/test_events.py
@@ -1838,7 +1838,7 @@ class NormalActionsTest(BaseAction):
         )
         events = self.verify_action(
             lambda: do_change_user_group_permission_setting(
-                backend, "can_mention_groups", moderators_group, acting_user=None
+                backend, "can_mention_groups", set({moderators_group.id}), acting_user=None
             )
         )
         check_user_group_update("events[0]", events[0], "can_mention_groups")